requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
gymnasium>=0.29.0
stable-baselines3[extra]>=2.2.0
streamlit>=1.28.0
plotly>=5.15.0
pyvista>=0.42.0
trimesh>=4.0.0
pyyaml>=6.0
docker>=6.1.0
shapely>=2.0.0
networkx>=3.1
torch>=2.0.0
tqdm>=4.65.0
